+++ Logger.h
@@ -5,6 +5,7 @@
 #include <string>
 #ifndef LOGGER_H
 #define LOGGER_H
+#pragma once
