+++ ThreatLevel.h
@@ -1,0 +1,13 @@
+//
+// Created by MINALP1 on 5/26/2025.
+//
+
+#ifndef THREATLEVEL_H
+#define THREATLEVEL_H
+
+
+enum class ThreatLevel { Sunrise, Noon, Midnight };
+
+
+
+#endif //THREATLEVEL_H
